src/panoseti_grpc/ublox_control/initialize/requirements.txt
json5==0.12.0
pyserial==3.5
pyubx2==1.2.55
